visual changes

+++ src/components/DashboardItem.js
@@ -6,21 +6,46 @@ class DashboardItem extends React.Component {
     constructor(props) {
         super(props);
 
-        this.edit = this.edit.bind(this)
-        this.delete = this.delete.bind(this)
-        this.rename = this.rename.bind(this)
+        this.edit = this.edit.bind(this);
+        this.delete = this.delete.bind(this);
+        this.rename = this.rename.bind(this);
+
+        this.closeOptions = this.closeOptions.bind(this);
+        this.openOptions = this.openOptions.bind(this);
+
+        this.handleEscape = this.handleEscape.bind(this);
+        this.handleClick = this.handleClick.bind(this);
+
+        this.state = {
+            optionsOpen: false
+        };
+
+      
     }
 
+
     delete() {
-        this.props.deleteDeck(this.props.id)
+        const decision = confirm(`Are you sure you want to delete deck: ${this.props.name}?`);
+        if (decision) {
+            this.closeOptions();
+            this.props.deleteDeck(this.props.id);
+        } else {
+            this.closeOptions();
+        }
+
     }
     rename() {
-        const name = prompt(`Enter new name for ${this.props.name}: `);
+        const name = prompt(`Enter new name for ${this.props.name}: `, this.props.name);
         if (name) {
-            this.props.renameDeck(this.props.id, name)
+            this.props.renameDeck(this.props.id, name);
+            this.closeOptions();
+        } else {
+            this.closeOptions();
         }
+
     }
-    edit() {       
+    edit() {    
+        this.closeOptions();
         this.props.history.push({
             pathname: '/edit/' + this.props.id,
             state: {
@@ -28,19 +53,67 @@ class DashboardItem extends React.Component {
             }
         });
     }
-    render() {
-        return (
-            <li className='dashboarditem'>
 
-                <div className="dashboarditem-label">{this.props.name}</div> 
-                
-                <div className="dashboard-buttons">
-                    <div className="dashboard-button dashboard-button-rename" onClick={this.rename}>Rename</div>
-                    <div className="dashboard-button dashboard-button-edit" onClick={this.edit}>Edit</div>
-                    <div className="dashboard-button dashboard-button-delete" onClick={this.delete}>Delete</div>
-                </div>
-            </li>
-        );
+    handleEscape(e) {
+        // console.log('event', e)
+        if (e.key === 'Escape') {
+            this.closeOptions();
+        }
+    }
+    handleClick(e) {
+        // console.log('click', e)
+        // console.log(this)
+        if (this.node.contains(e.target)) {
+            // console.log('inside');
+            return;
+        }
+        // console.log('outside')
+        this.closeOptions();
+
+    }
+    closeOptions() {
+        this.setState({
+            optionsOpen: false
+        });
+        
+        document.removeEventListener('keydown', this.handleEscape, false);
+        document.removeEventListener('click', this.handleClick, false);
+    }
+    openOptions() {
+        this.setState({
+            optionsOpen: true
+        });
+
+        document.addEventListener('keydown', this.handleEscape, false);
+        document.addEventListener('click', this.handleClick, false);
+    }
+    render() {
+
+        if (this.state.optionsOpen) {
+            return (
+                <li className='dashboard-item'
+                    ref={node => { this.node = node; }}
+                >
+                    <div className="dashboard-control">
+                        <div className="dashboard-control-button dashboard-control-button-edit"
+                        onClick={this.edit} >edit</div>
+                        <div className="dashboard-control-button dashboard-control-button-rename"
+                        onClick={this.rename} >rename</div>
+                        <div className="dashboard-control-button dashboard-control-button-delete"
+                        onClick={this.delete} >delete</div>
+                        <div className="dashboard-control-button dashboard-control-button-close"
+                        onClick={this.closeOptions}>x</div>
+                    </div>
+                </li>
+            );
+        } else {
+            return (
+                <li className='dashboard-item'>
+                    <div className="dashboard-label">{this.props.name}</div> 
+                    <div className="dashboard-actions" onClick={this.openOptions}>Actions</div>
+                </li>
+            );
+        }
     }
 }
